fixup: move tests over to state package
